% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{im2ansi}
\alias{im2ansi}
\title{Convert an image to an ANSI string}
\usage{
im2ansi(im, width = 80, font_aspect = 0.45, full_colour = FALSE)
}
\arguments{
\item{im}{magick image}

\item{width}{text output width. default: 80}

\item{font_aspect}{ratio of width to height for the output font. default: 0.45}

\item{full_colour}{use 24bit colour codes. default: FALSE. Only supported by
some terminals}
}
\value{
Character string with ANSI colours
}
\description{
Convert an image to an ANSI string
}
\examples{
\dontrun{
im <- image_read(system.file('img', 'Rlogo.png', package = 'png'))
cat(im2ansi(im, width = 80))
}
}
