% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{im2char}
\alias{im2char}
\title{Convert an image to a matrix of ANSI chars}
\usage{
im2char(im, width = 80, font_aspect = 0.45)
}
\arguments{
\item{im}{magick image}

\item{width}{text output width. default: 80}

\item{font_aspect}{ratio of width to height for the output font. default: 0.5}
}
\value{
Character string with ANSI colours
}
\description{
Convert an image to a matrix of ANSI chars
}
\examples{
\dontrun{
im <- image_read(system.file('img', 'Rlogo.png', package = 'png'))
mat <- im2char(im)
}
}
